requirements.txt
# requirements.txt
requests==2.31.0
pandas>=2.2.0
sqlalchemy>=2.0.20
python-dotenv>=1.0.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
nltk>=3.8.1
scikit-learn>=1.3.0
plotly>=5.17.0
schedule>=1.2.0
